backend/requirements.txt
fastapi==0.104.1
uvicorn[all]
python-multipart==0.0.6
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.26.2
